refactor(areas-seeding): moved wrongly place migration into workflows/seed directory
Took 4 minutes
